+++ src/fidesctl/cli/commands/core.py
@@ -7,12 +7,13 @@ from fidesctl.cli.options import (
     manifests_dir_argument,
     verbose_flag,
 )
-from fidesctl.cli.utils import pretty_echo, print_divider, with_analytics
+from fidesctl.cli.utils import echo_red, pretty_echo, print_divider, with_analytics
 from fidesctl.core import apply as _apply
 from fidesctl.core import audit as _audit
 from fidesctl.core import evaluate as _evaluate
 from fidesctl.core import parse as _parse
 from fidesctl.core import sync as _sync
+from fidesctl.core.utils import git_is_dirty
 
 
 @click.command()
@@ -141,11 +142,18 @@ def parse(ctx: click.Context, manifests_dir: str, verbose: bool = False) -> None
 def sync(ctx: click.Context, manifests_dir: str) -> None:
     """
     Update local resource files by their fides_key to match their server versions.
+
+    Aborts the sync if there are unstaged or untracked files in the manifests dir.
     """
 
     config = ctx.obj["CONFIG"]
     # Do this to validate the manifests since they won't get parsed during the sync process
     _parse.parse(manifests_dir)
+    if git_is_dirty(manifests_dir):
+        echo_red(
+            f"There are unstaged changes in your manifest directory: '{manifests_dir}' \nAborting sync!"
+        )
+        raise SystemExit(1)
     _sync.sync(
         url=config.cli.server_url,
         manifests_dir=manifests_dir,